Initial commit of main functionality
The entry point is `stygian.html`. This loads the external depedencies and
included js files. This is not optimal for page-load speeds, but it makes it
easier to read the source.

The generated map of the library is based on
[`vis.js`](https://visjs.github.io/vis-network/docs/network/).

All text is attributed to Emmy "Cavegirl" Allen, the original creator of the
Stygian Library, which can be bought
[here](https://www.drivethrurpg.com/product/257113/The-Stygian-Library).